usm: Remove debugger_windows.go
Removing unnecessary windows implementation
