+++ pkg/network/encoding/marshal/usm_http2_test.go
@@ -3,10 +3,11 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build linux && linux_bpf
+
 package marshal
 
 import (
-	"runtime"
 	"testing"
 
 	model "github.com/DataDog/agent-payload/v5/process"
@@ -35,9 +36,6 @@ type HTTP2Suite struct {
 }
 
 func TestHTTP2Stats(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("the feature is only supported on linux.")
-	}
 	suite.Run(t, &HTTP2Suite{})
 }
 
@@ -86,9 +84,11 @@ func (s *HTTP2Suite) TestFormatHTTP2Stats() {
 				}},
 			},
 		},
-		HTTP2: map[http.Key]*http.RequestStats{
-			httpKey1: http2Stats1,
-			httpKey2: http2Stats2,
+		USMData: network.USMProtocolsData{
+			HTTP2: map[http.Key]*http.RequestStats{
+				httpKey1: http2Stats1,
+				httpKey2: http2Stats2,
+			},
 		},
 	}
 	out := &model.HTTP2Aggregations{
@@ -113,7 +113,7 @@ func (s *HTTP2Suite) TestFormatHTTP2Stats() {
 		out.EndpointAggregations[1].StatsByStatusCode[int32(statusCode)] = &model.HTTPStats_Data{Count: 1, FirstLatencySample: 20, Latencies: nil}
 	}
 
-	http2Encoder := newHTTP2Encoder(in.HTTP2)
+	http2Encoder := newHTTP2Encoder(in.USMData.HTTP2)
 	aggregations, tags, _ := getHTTP2Aggregations(t, http2Encoder, in.Conns[0])
 
 	require.NotNil(t, aggregations)
@@ -163,11 +163,13 @@ func (s *HTTP2Suite) TestFormatHTTP2StatsByPath() {
 				}},
 			},
 		},
-		HTTP2: map[http.Key]*http.RequestStats{
-			key: http2ReqStats,
+		USMData: network.USMProtocolsData{
+			HTTP2: map[http.Key]*http.RequestStats{
+				key: http2ReqStats,
+			},
 		},
 	}
-	http2Encoder := newHTTP2Encoder(payload.HTTP2)
+	http2Encoder := newHTTP2Encoder(payload.USMData.HTTP2)
 	http2Aggregations, tags, _ := getHTTP2Aggregations(t, http2Encoder, payload.Conns[0])
 
 	require.NotNil(t, http2Aggregations)
@@ -233,12 +235,14 @@ func (s *HTTP2Suite) TestHTTP2IDCollisionRegression() {
 		BufferedData: network.BufferedData{
 			Conns: connections,
 		},
-		HTTP2: map[http.Key]*http.RequestStats{
-			httpKey: http2Stats,
+		USMData: network.USMProtocolsData{
+			HTTP2: map[http.Key]*http.RequestStats{
+				httpKey: http2Stats,
+			},
 		},
 	}
 
-	http2Encoder := newHTTP2Encoder(in.HTTP2)
+	http2Encoder := newHTTP2Encoder(in.USMData.HTTP2)
 
 	// assert that the first connection matching the HTTP2 data will get
 	// back a non-nil result
@@ -296,30 +300,13 @@ func (s *HTTP2Suite) TestHTTP2LocalhostScenario() {
 		BufferedData: network.BufferedData{
 			Conns: connections,
 		},
-		HTTP2: map[http.Key]*http.RequestStats{
-			httpKey: http2Stats,
+		USMData: network.USMProtocolsData{
+			HTTP2: map[http.Key]*http.RequestStats{
+				httpKey: http2Stats,
+			},
 		},
 	}
-	if runtime.GOOS == "windows" {
-		/*
-		 * on Windows, there are separate http transactions for
-		 * each side of the connection.  And they're kept separate,
-		 * and keyed separately.  Address this condition until the
-		 * platforms are resynced
-		 */
-		httpKeyWin := http.NewKey(
-			util.AddressFromString("127.0.0.1"),
-			util.AddressFromString("127.0.0.1"),
-			serverport,
-			cliport,
-			[]byte("/"),
-			true,
-			http.MethodGet,
-		)
-
-		in.HTTP2[httpKeyWin] = http2Stats
-	}
-	http2Encoder := newHTTP2Encoder(in.HTTP2)
+	http2Encoder := newHTTP2Encoder(in.USMData.HTTP2)
 
 	// assert that both ends (client:server, server:client) of the connection
 	// will have HTTP2 stats
