+++ pkg/network/encoding/marshal/usm_kafka_test.go
@@ -3,11 +3,12 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build linux && linux_bpf
+
 package marshal
 
 import (
 	"fmt"
-	"runtime"
 	"testing"
 
 	"github.com/gogo/protobuf/proto"
@@ -21,12 +22,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/kafka"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 )
-
-func skipIfNotLinux(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("the feature is only supported on linux.")
-	}
-}
 
 const (
 	clientPort  = uint16(1234)
@@ -51,7 +46,6 @@ type KafkaSuite struct {
 }
 
 func TestKafkaStats(t *testing.T) {
-	skipIfNotLinux(t)
 	suite.Run(t, &KafkaSuite{})
 }
 
@@ -83,17 +77,19 @@ func (s *KafkaSuite) TestFormatKafkaStats() {
 				defaultConnection,
 			},
 		},
-		Kafka: map[kafka.Key]*kafka.RequestStats{
-			kafkaKey1: {
-				ErrorCodeToStat: map[int32]*kafka.RequestStat{
-					0: {Count: 10},
-					1: {Count: 2},
+		USMData: network.USMProtocolsData{
+			Kafka: map[kafka.Key]*kafka.RequestStats{
+				kafkaKey1: {
+					ErrorCodeToStat: map[int32]*kafka.RequestStat{
+						0: {Count: 10},
+						1: {Count: 2},
+					},
 				},
-			},
-			kafkaKey2: {
-				ErrorCodeToStat: map[int32]*kafka.RequestStat{
-					0:  {Count: 2},
-					10: {Count: 5},
+				kafkaKey2: {
+					ErrorCodeToStat: map[int32]*kafka.RequestStat{
+						0:  {Count: 2},
+						10: {Count: 5},
+					},
 				},
 			},
 		},
@@ -119,7 +115,7 @@ func (s *KafkaSuite) TestFormatKafkaStats() {
 		},
 	}
 
-	encoder := newKafkaEncoder(in.Kafka)
+	encoder := newKafkaEncoder(in.USMData.Kafka)
 	t.Cleanup(encoder.Close)
 
 	aggregations := getKafkaAggregations(t, encoder, in.Conns[0])
@@ -162,16 +158,18 @@ func (s *KafkaSuite) TestKafkaIDCollisionRegression() {
 		BufferedData: network.BufferedData{
 			Conns: connections,
 		},
-		Kafka: map[kafka.Key]*kafka.RequestStats{
-			kafkaKey: {
-				ErrorCodeToStat: map[int32]*kafka.RequestStat{
-					0: {Count: 10},
+		USMData: network.USMProtocolsData{
+			Kafka: map[kafka.Key]*kafka.RequestStats{
+				kafkaKey: {
+					ErrorCodeToStat: map[int32]*kafka.RequestStat{
+						0: {Count: 10},
+					},
 				},
 			},
 		},
 	}
 
-	encoder := newKafkaEncoder(in.Kafka)
+	encoder := newKafkaEncoder(in.USMData.Kafka)
 	t.Cleanup(encoder.Close)
 	aggregations := getKafkaAggregations(t, encoder, in.Conns[0])
 
@@ -223,16 +221,18 @@ func (s *KafkaSuite) TestKafkaLocalhostScenario() {
 		BufferedData: network.BufferedData{
 			Conns: connections,
 		},
-		Kafka: map[kafka.Key]*kafka.RequestStats{
-			kafkaKey: {
-				ErrorCodeToStat: map[int32]*kafka.RequestStat{
-					0: {Count: 10},
+		USMData: network.USMProtocolsData{
+			Kafka: map[kafka.Key]*kafka.RequestStats{
+				kafkaKey: {
+					ErrorCodeToStat: map[int32]*kafka.RequestStat{
+						0: {Count: 10},
+					},
 				},
 			},
 		},
 	}
 
-	encoder := newKafkaEncoder(in.Kafka)
+	encoder := newKafkaEncoder(in.USMData.Kafka)
 	t.Cleanup(encoder.Close)
 
 	// assert that both ends (client:server, server:client) of the connection
@@ -265,7 +265,9 @@ func generateBenchMarkPayloadKafka(entries uint16) network.Connections {
 		BufferedData: network.BufferedData{
 			Conns: make([]network.ConnectionStats, 1),
 		},
-		Kafka: map[kafka.Key]*kafka.RequestStats{},
+		USMData: network.USMProtocolsData{
+			Kafka: map[kafka.Key]*kafka.RequestStats{},
+		},
 	}
 
 	payload.Conns[0].Dest = localhost
@@ -274,7 +276,7 @@ func generateBenchMarkPayloadKafka(entries uint16) network.Connections {
 	payload.Conns[0].SPort = 1112
 
 	for index := uint16(0); index < entries; index++ {
-		payload.Kafka[kafka.NewKey(
+		payload.USMData.Kafka[kafka.NewKey(
 			localhost,
 			localhost,
 			1112,
@@ -298,7 +300,7 @@ func commonBenchmarkKafkaEncoder(b *testing.B, entries uint16) {
 	b.ReportAllocs()
 	var h *kafkaEncoder
 	for i := 0; i < b.N; i++ {
-		h = newKafkaEncoder(payload.Kafka)
+		h = newKafkaEncoder(payload.USMData.Kafka)
 		streamer.Reset()
 		h.EncodeConnection(payload.Conns[0], a)
 		h.Close()
