+++ pkg/network/encoding/marshal/usm_lookup.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows
+//go:build linux && linux_bpf
 
 package marshal
