+++ pkg/network/protocols/redis/debugging/debugging.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux_bpf
+//go:build linux_bpf && linux
 
 // Package debugging provides debug-friendly representation of internal data structures
 package debugging
