+++ pkg/network/tracer/tracer.go
@@ -458,11 +458,7 @@ func (t *Tracer) GetActiveConnections(clientID string) (*network.Connections, fu
 	buffer.ConnectionBuffer.Assign(delta.Conns)
 	conns := network.NewConnections(buffer)
 	conns.DNS = t.reverseDNS.Resolve(ips)
-	conns.HTTP = delta.HTTP
-	conns.HTTP2 = delta.HTTP2
-	conns.Kafka = delta.Kafka
-	conns.Postgres = delta.Postgres
-	conns.Redis = delta.Redis
+	conns.USMData = delta.USMData
 	conns.ConnTelemetry = t.state.GetTelemetryDelta(clientID, t.getConnTelemetry(len(active)))
 	conns.CompilationTelemetryByAsset = t.getRuntimeCompilationTelemetry()
 	conns.KernelHeaderFetchResult = int32(headers.HeaderProvider.GetResult())
